+++ app/model/recordings.js
@@ -1660,6 +1660,15 @@ var Recordings = {
         return queryResult;
     },
 
+    writeExportParams: function(projection, filters, userId, userEmail) {
+        return Q.ninvoke(joi, 'validate', projection, Recordings.SCHEMAS.exportProjections)
+            .then((projection_parameters) => {
+                const q = `INSERT INTO recordings_export_parameters (project_id, user_id, user_email, projection_parameters, filters, created_at, processed_at, error)
+                VALUES(${filters.project_id}, ${userId}, '${userEmail}', '${JSON.stringify(projection_parameters)}', '${JSON.stringify(filters)}', NOW(), null, null)`
+                return dbpool.query(q)
+            })
+    },
+
     groupedDetections: async function(projection, filters){
         let summaryBuilders = [];
         return Q.ninvoke(joi, 'validate', projection, Recordings.SCHEMAS.exportProjections)
@@ -1680,7 +1689,7 @@ var Recordings = {
                             let index = 0;
                             for (let c in classesArray) {
                                 let builder = new SQLBuilder();
-                                builder.setOrderBy(filters.sortBy || 'site', !filters.sortRev);
+                                builder.setOrderBy('s.name');
                                 summaryBuilders.push(builder);
                                 classesArray[c].forEach(function(cls, idx){
                                     index++;
@@ -1715,7 +1724,6 @@ var Recordings = {
                     // Get rows for the detections grouped by site.
                     if (projection.grouped && projection.grouped === 'site') {
                         summaryBuilders[c].addProjection('s.name as site');
-                        // summaryBuilders[c].addConstraint('s.site_id=22489');
                     }
                     // Get rows for the detections grouped by date.
                     if (projection.grouped && projection.grouped === 'date') {
@@ -1732,7 +1740,7 @@ var Recordings = {
             }).then(async function() {
                 let results = [];
                 for (let builder in summaryBuilders) {
-                    let queryResult = await (projection.grouped ? dbpool.query : dbpool.streamQuery)({
+                    let queryResult = await dbpool.query({
                         sql: summaryBuilders[builder].getSQL(),
                         typeCast: sqlutil.parseUtcDatetime,
                     })
@@ -1743,7 +1751,7 @@ var Recordings = {
 
     },
 
-    exportRecordingData: function(projection, filters){
+    exportRecordingData: async function(projection, filters) {
         let summaryBuilders = [];
         return Q.ninvoke(joi, 'validate', projection, Recordings.SCHEMAS.exportProjections)
             .then(async (all) => {
@@ -1795,21 +1803,7 @@ var Recordings = {
                     summaryBuilders.push(builder);
                 }
                 for (let c in summaryBuilders) {
-                    // Get rows for the detections grouped by site.
-                    if (projection.grouped && projection.grouped === 'site') {
-                        summaryBuilders[c].addProjection('s.name as site');
-                    }
-                    // Get rows for the detections grouped by date.
-                    if (projection.grouped && projection.grouped === 'date') {
-                        summaryBuilders[c].addProjection('DATE_FORMAT(r.datetime, "%Y/%m/%d") as date');
-                        summaryBuilders[c].setOrderBy('r.datetime');
-                    }
-                    // Get rows for the detections grouped by hour.
-                    if (projection.grouped && projection.grouped === 'hour') {
-                        summaryBuilders[c].addProjection('DATE_FORMAT(r.datetime, "%H") as hour');
-                        summaryBuilders[c].setOrderBy('r.datetime');
-                    }
-                    if (projection_parameters.recording && !projection.grouped) {
+                    if (projection_parameters.recording) {
                         var recParamMap = {
                             'filename' : "SUBSTRING_INDEX(r.uri,'/',-1) as filename",
                             'site' : 's.name as site',
@@ -1830,7 +1824,7 @@ var Recordings = {
                             summaryBuilders[c].setOrderBy('r.datetime');
                         }
                     }
-                    if (projection_parameters.classification && !projection.grouped) {
+                    if (projection_parameters.classification) {
                         promises.push(models.classifications.getFor({id:projection_parameters.classification, showModel:true}).then(function(classifications){
                             classifications.forEach(function(classification, idx){
                                 var clsid = "p_CR_" + idx;
@@ -1846,7 +1840,7 @@ var Recordings = {
                             });
                         }));
                     }
-                    if (projection_parameters.soundscapeComposition && !projection.grouped) {
+                    if (projection_parameters.soundscapeComposition) {
                         promises.push(models.SoundscapeComposition.getClassesFor({id:projection_parameters.soundscapeComposition}).then(function(classes){
                             classes.forEach(function(cls, idx){
                                 var clsid = "p_SCC_" + idx;
@@ -1859,7 +1853,7 @@ var Recordings = {
                             });
                         }));
                     }
-                    if(projection_parameters.tag && !projection.grouped){
+                    if(projection_parameters.tag){
                         promises.push(models.tags.getFor({id:projection_parameters.tag}).then(function(tags){
                             tags.forEach(function(tag, idx){
                                 var prtid = "p_RT_" + idx;
@@ -1873,19 +1867,29 @@ var Recordings = {
             }).then(async function() {
                 let results = [];
                 for (let builder in summaryBuilders) {
-                    let queryResult = await (projection.grouped ? dbpool.query : dbpool.streamQuery)({
+                    let queryResult = await dbpool.streamQuery({
                         sql: summaryBuilders[builder].getSQL(),
                         typeCast: sqlutil.parseUtcDatetime,
                     })
-                    if (projection.grouped) {
-                        results = results.concat([...new Set(queryResult)]);
-                    }
-                    else {
-                        results = [...new Set(queryResult)];
-                    }
+                    results = [...new Set(queryResult)];
                 }
-                return projection.grouped? results : Q.all(results);
+                return Q.all(results);
             })
+    },
+
+    exportOccupancyModels: async function(projection, filters){
+        let query = `SELECT S.name as site, DATE_FORMAT(R.datetime, "%Y/%m/%d") as date, SUM(rv.present=1 OR rv.present_review>0 OR rv.present_aed>0) as count
+            FROM recordings R
+            JOIN sites S ON S.site_id = R.site_id
+            LEFT JOIN project_imported_sites AS pis ON S.site_id = pis.site_id AND pis.project_id = ${filters.project_id}
+            LEFT JOIN recording_validations AS rv ON R.recording_id = rv.recording_id
+            WHERE rv.species_id = ${projection.species} AND (S.project_id = ${filters.project_id} OR pis.project_id = ${filters.project_id}) AND (rv.present_review>0 OR rv.present_aed>0 OR rv.present is not null)
+            GROUP BY S.name, YEAR(R.datetime), MONTH(R.datetime), DAY(R.datetime) ORDER BY R.datetime ASC`;
+        let queryResult = await dbpool.query({
+            sql: query,
+            typeCast: sqlutil.parseUtcDatetime,
+        })
+        return queryResult;
     },
 
     countProjectSpecies: function(project_id) {